refactor(core,preact): Moved FetchResult type from core to preact workspace and added usage example of ValidateStatusMiddleware
