+++ packages/pretch-preact/hooks/use-fetch.ts
@@ -1,5 +1,6 @@
 import { useSignal, useSignalEffect } from "@preact/signals";
-import { buildFetch, type Enhancer, type FetchResult } from "@pretch/core";
+import { buildFetch, type Enhancer } from "@pretch/core";
+import type {FetchResult } from "@/types.ts";
 
 export function useFetch<T>(
   url: string | URL,
