+++ packages/pretch-preact/mod.ts
@@ -1,1 +1,2 @@
 export { useFetch } from "@/hooks/use-fetch.ts";
+export * from "@/types.ts";
